+++ src/media/ofxVlc4Media.cpp
@@ -4303,6 +4303,101 @@ void ofxVlc4::navigate(NavigationMode mode) {
 	mediaComponent->navigate(mode);
 }
 
+bool ofxVlc4::executePlayerCommand(PlayerCommand command) {
+	switch (command) {
+	case PlayerCommand::PlayPause:
+		if (isPlaying()) {
+			pause();
+		} else {
+			play();
+		}
+		return true;
+	case PlayerCommand::Play:
+		play();
+		return true;
+	case PlayerCommand::Pause:
+		pause();
+		return true;
+	case PlayerCommand::Stop:
+		stop();
+		return true;
+	case PlayerCommand::NextItem:
+		nextMediaListItem();
+		return true;
+	case PlayerCommand::PreviousItem:
+		previousMediaListItem();
+		return true;
+	case PlayerCommand::SeekForwardSmall:
+		setTime(getTime() + 5000);
+		return true;
+	case PlayerCommand::SeekBackwardSmall:
+		setTime(getTime() - 5000);
+		return true;
+	case PlayerCommand::SeekForwardLarge:
+		setTime(getTime() + 30000);
+		return true;
+	case PlayerCommand::SeekBackwardLarge:
+		setTime(getTime() - 30000);
+		return true;
+	case PlayerCommand::VolumeUp:
+		setVolume(getVolume() + 5);
+		return true;
+	case PlayerCommand::VolumeDown:
+		setVolume(getVolume() - 5);
+		return true;
+	case PlayerCommand::ToggleMute:
+		toggleMute();
+		return true;
+	case PlayerCommand::NextFrame:
+		nextFrame();
+		return true;
+	case PlayerCommand::PreviousChapter:
+		previousChapter();
+		return true;
+	case PlayerCommand::NextChapter:
+		nextChapter();
+		return true;
+	case PlayerCommand::MenuActivate:
+		navigate(NavigationMode::Activate);
+		return true;
+	case PlayerCommand::MenuUp:
+		navigate(NavigationMode::Up);
+		return true;
+	case PlayerCommand::MenuDown:
+		navigate(NavigationMode::Down);
+		return true;
+	case PlayerCommand::MenuLeft:
+		navigate(NavigationMode::Left);
+		return true;
+	case PlayerCommand::MenuRight:
+		navigate(NavigationMode::Right);
+		return true;
+	case PlayerCommand::MenuPopup:
+		navigate(NavigationMode::Popup);
+		return true;
+	case PlayerCommand::TeletextRed:
+		sendTeletextKey(TeletextKey::Red);
+		return true;
+	case PlayerCommand::TeletextGreen:
+		sendTeletextKey(TeletextKey::Green);
+		return true;
+	case PlayerCommand::TeletextYellow:
+		sendTeletextKey(TeletextKey::Yellow);
+		return true;
+	case PlayerCommand::TeletextBlue:
+		sendTeletextKey(TeletextKey::Blue);
+		return true;
+	case PlayerCommand::TeletextIndex:
+		sendTeletextKey(TeletextKey::Index);
+		return true;
+	case PlayerCommand::ToggleTeletextTransparency:
+		setTeletextTransparencyEnabled(!isTeletextTransparencyEnabled());
+		return true;
+	}
+
+	return false;
+}
+
 std::vector<ofxVlc4::MediaTrackInfo> ofxVlc4::MediaComponent::getTrackInfos(libvlc_track_type_t type) const {
 	std::vector<MediaTrackInfo> trackInfos;
 	libvlc_media_player_t * player = owner.sessionPlayer();
@@ -4473,4 +4568,3 @@ bool ofxVlc4::selectSubtitleTrackById(const std::string & trackId) {
 void ofxVlc4::clearCurrentMedia(bool clearVideoResources) {
 	mediaComponent->clearCurrentMedia(clearVideoResources);
 }
-